requirements.txt
aiofiles==24.1.0
aioodbc==0.5.0
aioredis==2.0.1
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
azure-cognitiveservices-language-textanalytics==0.2.2
azure-common==1.1.28
azure-core==1.35.1
azure-identity==1.25.0
azure-keyvault-secrets==4.10.0
azure-mgmt-core==1.6.0
azure-storage-blob==12.26.0
bcrypt==4.3.0
black==25.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==46.0.1
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.117.1
flake8==7.3.0
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
isodate==0.7.2
isort==6.0.1
jiter==0.11.0
Mako==1.3.10
MarkupSafe==3.0.2
mccabe==0.7.0
motor==3.7.1
msal==1.33.0
msal-extensions==1.3.1
msrest==0.7.1
mypy==1.18.2
mypy_extensions==1.1.0
oauthlib==3.3.1
openai==1.108.1
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.15
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.1
pyodbc==5.3.0
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytokens==0.1.10
PyYAML==6.0.2
redis==6.4.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
structlog==25.4.0
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.36.0
watchfiles==1.1.0
websockets==15.0.1
fastapi==0.117.1
uvicorn[standard]==0.40.0
pydantic[email]
python-dotenv
bcrypt
python-jose
pydantic[email]
bcrypt==4.3.0
python-jose[cryptography]
python-jose[cryptography]==3.5.0